remove login page
