+++ src/pages/Home.tsx
@@ -1,9 +1,23 @@
 import React from "react"
+import { Button } from "@material-ui/core"
+import { hostClient } from "./const"
 
 export const Home = () => {
+  const redirectUri = encodeURIComponent(`${hostClient}/spotify-callback`)
+  const scope = encodeURIComponent(
+    "user-top-read user-read-private user-read-email"
+  )
+  const clientId = "4ff85ab0e8334991bbfdbb920765ec70"
+
   return (
     <>
-      <h2>ホーム</h2>
+      <Button
+        onClick={() => {
+          window.location.href = `https://accounts.spotify.com/authorize?response_type=code&client_id=${clientId}&scope=${scope}&redirect_uri=${redirectUri}`
+        }}
+      >
+        Spoitfyで登録
+      </Button>
     </>
   )
 }
